+++ marketscrape/frontend/src/components/ListingInputCard.vue
